Finished Route class time and distance calculations. Added method in MainActivity to show and hide bottom bar.

+++ app/src/main/java/edu/utdallas/whoosh/api/IRoute.java
@@ -34,7 +34,7 @@ public interface IRoute {
     /**
      * the estimated length of time in minutes to complete this route, given its {@link #getType() mode of transit}
      */
-    int getTimeInMinutes();
+    float getTimeInMinutes();
 
     /**
      * the calculated distance in feet of this route
